Separate KeyValueStore module on its own file
