+++ lib/todo/key_value_store.ex
@@ -1,0 +1,29 @@
+defmodule KeyValueStore do
+  use GenServer
+
+  def init(_) do
+    {:ok, %{}}
+  end
+
+  def start do
+    GenServer.start(KeyValueStore, nil)
+  end
+
+  def put(pid, key, value) do
+    GenServer.cast(pid, {:put, key, value})
+  end
+
+  def get(pid, key) do
+    GenServer.call(pid, {:get, key})
+  end
+
+  # Handles the put request
+  def handle_cast({:put, key, value}, state) do
+    {:noreply, Map.put(state, key, value)}
+  end
+
+  # Handles the get request
+  def handle_call({:get, key}, _, state) do
+    {:reply, Map.get(state, key), state}
+  end
+end
