+++ lib/todo/server_process.ex
@@ -45,33 +45,3 @@ defmodule ServerProcess do
     send(server_pid, {:cast, request})
   end
 end
-
-defmodule KeyValueStore do
-  use GenServer
-
-  def init(_) do
-    {:ok, %{}}
-  end
-
-  def start do
-    GenServer.start(KeyValueStore, nil)
-  end
-
-  def put(pid, key, value) do
-    GenServer.cast(pid, {:put, key, value})
-  end
-
-  def get(pid, key) do
-    GenServer.call(pid, {:get, key})
-  end
-
-  # Handles the put request
-  def handle_cast({:put, key, value}, state) do
-    {:noreply, Map.put(state, key, value)}
-  end
-
-  # Handles the get request
-  def handle_call({:get, key}, _, state) do
-    {:reply, Map.get(state, key), state}
-  end
-end
